fixing flutter analyze
