+++ lib/components/panel/species_panel.dart
@@ -2,24 +2,42 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:sliding_up_panel/sliding_up_panel.dart';
 import 'package:smartflore/bloc/map/map_bloc.dart';
-import 'package:smartflore/components/map/map_ui_widget.dart';
 import 'package:smartflore/components/map/map_widget.dart';
-import 'package:smartflore/components/panel/trail-list-panel.dart';
+import 'package:smartflore/components/panel/trail_list_panel.dart';
 
-class TrailsPanelWidget extends StatefulWidget {
+class SpeciesPanelWidget extends StatefulWidget {
   final bool isDraggable;
-  const TrailsPanelWidget({Key? key, this.isDraggable = true})
+  final Widget body;
+
+  const SpeciesPanelWidget(
+      {Key? key, this.isDraggable = true, required this.body})
       : super(key: key);
 
   @override
-  State<TrailsPanelWidget> createState() => _TrailsPanelWidgetState();
+  State<SpeciesPanelWidget> createState() => _SpeciesPanelWidgetState();
 }
 
-class _TrailsPanelWidgetState extends State<TrailsPanelWidget> {
+class _SpeciesPanelWidgetState extends State<SpeciesPanelWidget>
+    with SingleTickerProviderStateMixin {
   final PanelController _panelController = PanelController();
   bool isPanelOpened = false;
   bool isPanelMoving = false;
-  bool showTrailPreview = false;
+  bool showMe = false;
+
+  late Animation<double> animation;
+  late AnimationController controller;
+
+  @override
+  void initState() {
+    super.initState();
+    controller = AnimationController(
+        vsync: this, duration: const Duration(milliseconds: 300));
+    animation = Tween<double>(begin: 0, end: 300).animate(
+        CurvedAnimation(parent: controller, curve: Curves.easeInOutCubic))
+      ..addListener(() {
+        setState(() {});
+      });
+  }
 
   void onPanUpdate(details) {
     // Swiping down
@@ -31,28 +49,26 @@ class _TrailsPanelWidgetState extends State<TrailsPanelWidget> {
     }
   }
 
+  void setShowMe(bool show) {
+    (show) ? controller.forward() : controller.reverse();
+    setState(() {
+      showMe = show;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    Color primary = Theme.of(context).colorScheme.primary;
     double screenH = MediaQuery.of(context).size.height;
     double screenW = MediaQuery.of(context).size.width;
     double bottomPadding = MediaQuery.of(context).padding.bottom / 4;
 
-    void trailPreviewPanel(bool show) {
-      if (show) _panelController.close();
-      setState(() {
-        showTrailPreview = show;
-      });
-    }
-
     return BlocListener<MapBloc, MapState>(
       listener: (context, state) {
         if (state is OnMapModeChanged) {
-          if (state.mapMode == MapMode.preview) {
-            trailPreviewPanel(true);
-          } else if (state.mapMode == MapMode.overview ||
-              state.mapMode == MapMode.trail) {
-            trailPreviewPanel(false);
+          if (state.mapMode == MapMode.trail) {
+            setShowMe(true);
+          } else {
+            setShowMe(false);
           }
         }
       },
@@ -60,23 +76,8 @@ class _TrailsPanelWidgetState extends State<TrailsPanelWidget> {
           backdropEnabled: true,
           backdropOpacity: 0,
           backdropTapClosesPanel: true,
-          parallaxEnabled: true,
-          parallaxOffset: .5,
           maxHeight: screenH * 0.8,
-          minHeight: 100 + bottomPadding,
-          isDraggable: widget.isDraggable,
-          onPanelOpened: () {
-            setState(() {
-              isPanelMoving = false;
-              isPanelOpened = true;
-            });
-          },
-          onPanelClosed: () {
-            setState(() {
-              isPanelMoving = false;
-              isPanelOpened = false;
-            });
-          },
+          minHeight: animation.value,
           controller: _panelController,
           header: SizedBox(
             width: screenW,
@@ -111,26 +112,7 @@ class _TrailsPanelWidgetState extends State<TrailsPanelWidget> {
                       ),
                     ),
                     const SizedBox(height: 8),
-                    Container(
-                      height: 46,
-                      decoration: BoxDecoration(
-                          border: Border.all(
-                              color: Theme.of(context).colorScheme.primary,
-                              width: 1),
-                          borderRadius:
-                              const BorderRadius.all(Radius.circular(6))),
-                      child: TabBar(
-                          isScrollable: false,
-                          unselectedLabelColor: primary,
-                          indicator: BoxDecoration(
-                              color: primary,
-                              borderRadius:
-                                  const BorderRadius.all(Radius.circular(4))),
-                          tabs: const [
-                            Tab(text: 'Tous les sentiers'),
-                            Tab(text: 'Mes sentiers')
-                          ]),
-                    ),
+                    Container(height: 46, color: Colors.red),
                   ],
                 ),
               ),
@@ -139,10 +121,7 @@ class _TrailsPanelWidgetState extends State<TrailsPanelWidget> {
           panelBuilder: (scrollController) => _buildSlidingPanel(
               scrollController: scrollController, bottomPadding: bottomPadding),
           borderRadius: const BorderRadius.vertical(top: Radius.circular(24)),
-          body: MapUIWidget(
-            bottomPadding: bottomPadding,
-            showTrailPreview: showTrailPreview,
-          )),
+          body: widget.body),
     );
   }
