+++ lib/components/panel/trail_list_panel.dart
